requirements.txt
causal-learn==0.1.2.3
pytorch>=1.11.0
pydot==1.4.2
scikit-learn==1.0.2
matplotlib>=3.5.1
pandas==1.4.1
tqdm==4.64.0
scipy==1.8.0
networkx==2.7.1